requirements.txt
APScheduler==3.11.2
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.2.1
TgCrypto==1.2.5
tzlocal==5.3.1
